feat: rename account
