+++ src/ledger/use-cases/virtual-account/create-virtual-account.use-case.spec.ts
@@ -5,9 +5,9 @@ import { CacheService } from '@/infrastructure/cache/services/cache.service';
 import {
   CreateVirtualAccountDto,
   VirtualAccountDto,
-} from '@/accounts/__defs__/accounts.dto';
-import { VirtualAccountsCacheKeys } from '@/accounts/utils';
-import { VirtualAccountService } from '@/accounts/services/virtual-account.service';
+} from '@/ledger/__defs__/accounts.dto';
+import { VirtualAccountsCacheKeys } from '@/ledger/utils';
+import { VirtualAccountService } from '@/ledger/services/virtual-account.service';
 
 jest.mock('@/infrastructure/cache/services/cache.service');
 jest.mock('@/accounts/services/virtual-account.service');
